Move test functions to test_groupings.py
Keeps main.py clear of clutter. These functions are rarely used.
