+++ main.py
@@ -74,63 +74,6 @@ returned_from_initialize = initialize(
       test_groupings.DEPEN_test_grouping02])
 
 
-
-
-# a temporary helper function to check each column against known categories
-def completeness_check(column_start, column_end, interest_in=None):
-    
-    check = {}
-    for var in PISA2012.columns[column_start:column_end]:
-        check[str(var)] = [str(var)]
-
-    completeness = initialize(
-        PISA2012.sample(500),
-        [category_definitions.KNOWN_CATEGORIES,
-          category_definitions.PREFERRED_NAMING,
-          check,
-          test_groupings.DEPEN_test_grouping01])[2]['indep_categories']
-
-    if interest_in:
-        for key in completeness:
-            if completeness!= interest_in:
-                print(key)
-
-    return completeness
-
-# check 100 at a time
-completeness_check(0, 20, None)
-
-
-def get_all_unique_short_categories(max_length=5,
-                                    column_start=None, column_end=None):
-    """
-    Pull sets of unique values from PISA2012 dataset for building 
-    collection of known categories.
-    """
-    found_unique_sets = []
-    for var in PISA2012.columns[column_start:column_end]:
-
-        # get unique_values, without nulls
-        unique_values = set({})
-        for unique_val in set(PISA2012[var].unique()):
-            if not pd.isnull(unique_val):
-                unique_values.add(unique_val.strip())
-
-        # check if already found, check length
-        if (unique_values not in found_unique_sets) and (1 < len(unique_values) < max_length):
-            # check for subsets
-            for past_match in found_unique_sets:
-                # skip if subset of existing set
-                if set(unique_values).issubset(past_match):
-                    unique_values = False
-                    break
-                # remove existing set if superset of existing set
-                elif set(past_match).issubset(unique_values):
-                    found_unique_sets.remove(past_match)
-            if unique_values:
-                found_unique_sets.append(unique_values)
-
-    return found_unique_sets
-
-all_short_uniques = get_all_unique_short_categories(20)
-
+# temporary testing, will delete later:
+test_groupings.completeness_check(PISA2012, 0, 20, None)
+all_short_uniques = test_groupings.get_all_unique_short_categories(PISA2012, 20)
