+++ test_groupings.py
@@ -47,3 +47,66 @@ INDEP_test_grouping99 = {
 DEPEN_test_grouping99 = {
     'grpname1': [],
     'grpname2': []}
+
+
+
+# A helper function to view a list of categories extracted from PISA2012
+# Can be used to create new category_definitons.
+def get_all_unique_short_categories(pisadf, max_length=5,
+                                    column_start=None, column_end=None):
+    """
+    Pull sets of unique values from PISA2012 dataset for building 
+    collection of known categories.
+    """
+    import pandas as pd
+    found_unique_sets = []
+    for var in pisadf.columns[column_start:column_end]:
+
+        # get unique_values, without nulls
+        unique_values = set({})
+        for unique_val in set(pisadf[var].unique()):
+            if not pd.isnull(unique_val):
+                unique_values.add(unique_val.strip())
+
+        # check if already found, check length
+        if (unique_values not in found_unique_sets) and (1 < len(unique_values) < max_length):
+            # check for subsets
+            for past_match in found_unique_sets:
+                # skip if subset of existing set
+                if set(unique_values).issubset(past_match):
+                    unique_values = False
+                    break
+                # remove existing set if superset of existing set
+                elif set(past_match).issubset(unique_values):
+                    found_unique_sets.remove(past_match)
+            if unique_values:
+                found_unique_sets.append(unique_values)
+
+    return found_unique_sets
+
+
+
+
+# A helper function to check each column against known categories
+def completeness_check(column_start, column_end, interest_in=None):
+    """
+    A helper function to check each column against known categories
+    """
+    check = {}
+    for var in PISA2012.columns[column_start:column_end]:
+        check[str(var)] = [str(var)]
+
+    completeness = initialize(
+        PISA2012.sample(500),
+        [category_definitions.KNOWN_CATEGORIES,
+          category_definitions.PREFERRED_NAMING,
+          check,
+          test_groupings.DEPEN_test_grouping01]
+        )[2]['indep_categories']
+
+    if interest_in:
+        for key in completeness:
+            if completeness!= interest_in:
+                print(key)
+
+    return completeness
